backend/flask_services/fairness_service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
aif360==0.5.0
fairlearn==0.9.0
numpy==1.24.3
pandas==2.1.3
psycopg2-binary==2.9.9
gunicorn==21.2.0

